refactor: move routing logic to nested components
